add: request dto de articulos de pedido, iniciado PedidoService

+++ src/main/java/com/api/multicode_reader/model/DetallePedido.java
@@ -24,5 +24,5 @@ public class DetallePedido {
     private Codigo codigo;
 
     @Column(name = "cantidad", nullable = false)
-    private int cantidad;
+    private double cantidad;
 }
